refactor(toast): split toast notification into generic Toast and Notification Visual
