+++ src/components/Notification/subcomponents/NotificationVisual/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './NotificationVisual';
